Incorporated PR comments (renamed GenderStrings to GenderValues and removed BooleanStrings enum
