+++ core/src/main/scala/com/salesforce/op/utils/stages/NameDetectUtils.scala
@@ -33,8 +33,8 @@ package com.salesforce.op.utils.stages
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.ser.std.StdSerializer
-import com.salesforce.op.features.types.NameStats.GenderStrings
-import com.salesforce.op.features.types.NameStats.GenderStrings._
+import com.salesforce.op.features.types.NameStats.GenderValues
+import com.salesforce.op.features.types.NameStats.GenderValues._
 import com.salesforce.op.features.types.Text
 import com.salesforce.op.stages.impl.feature.{GenderDetectStrategy, TextTokenizer}
 import com.salesforce.op.utils.json.{JsonLike, JsonUtils, SerDes}
@@ -101,7 +101,7 @@ private[op] trait NameDetectFun[T <: Text] extends Logging with NameDetectParams
     }
   }
 
-  private[op] def genderDictCheck(nameToCheckGenderOf: T#Value, genderDict: GenderDictionary): GenderStrings = {
+  private[op] def genderDictCheck(nameToCheckGenderOf: T#Value, genderDict: GenderDictionary): GenderValues = {
     nameToCheckGenderOf.flatMap(genderDict.get).map(
       probMale => if (probMale >= 0.5) Male else Female
     ).getOrElse(GenderNA)
@@ -112,7 +112,7 @@ private[op] trait NameDetectFun[T <: Text] extends Logging with NameDetectParams
     tokens: Seq[String],
     strategy: GenderDetectStrategy,
     genderDict: GenderDictionary
-  ): GenderStrings = {
+  ): GenderValues = {
     input match {
       case None => GenderNA
       case Some(text) =>
@@ -151,7 +151,7 @@ private[op] trait NameDetectFun[T <: Text] extends Logging with NameDetectParams
     genderDict: GenderDictionary
   ): Map[String, GenderStats] = {
     GenderDetectStrategies map { strategy: GenderDetectStrategy =>
-      val genderResult: GenderStrings = identifyGender(input, tokens, strategy, genderDict)
+      val genderResult: GenderValues = identifyGender(input, tokens, strategy, genderDict)
       implicit def booleanToInt(v: Boolean): Int = if (v) 1 else 0
       strategy.toString -> GenderStats(genderResult == Male, genderResult == Female, genderResult == GenderNA)
     } toMap
