+++ features/src/test/scala/com/salesforce/op/features/types/FeatureTypeSparkConverterTest.scala
@@ -84,10 +84,10 @@ class FeatureTypeSparkConverterTest
     Map("1" -> 3L, "2" -> 4L).toIntegralMap -> Map("1" -> 3L, "2" -> 4L),
     Map("1" -> "one", "2" -> "two").toTextMap -> Map("1" -> "one", "2" -> "two"),
     Map("1" -> Set("a", "b")).toMultiPickListMap -> Map("1" -> MWrappedArray.make(Array("a", "b"))),
-    Map(NameStats.Keys.IsName.toString -> NameStats.BooleanStrings.True.toString,
-      NameStats.Keys.Gender.toString -> NameStats.GenderStrings.Female.toString
-    ).toNameStats -> Map(NameStats.Keys.IsName.toString -> NameStats.BooleanStrings.True.toString,
-      NameStats.Keys.Gender.toString -> NameStats.GenderStrings.Female.toString
+    Map(NameStats.Keys.IsName.toString -> true.toString,
+      NameStats.Keys.Gender.toString -> NameStats.GenderValues.Female.toString
+    ).toNameStats -> Map(NameStats.Keys.IsName.toString -> true.toString,
+      NameStats.Keys.Gender.toString -> NameStats.GenderValues.Female.toString
     ),
     Map("1" -> Seq(1.0, 5.0, 6.0)).toGeolocationMap -> Map("1" -> MWrappedArray.make(Array(1.0, 5.0, 6.0)))
   )
